+++ Backend/Backend.Models/ProjectModels/ProjectOnlyNameModel.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Backend.Models.ProjectModels
+{
+    public class ProjectOnlyNameModel
+    {
+        public string Name { get; set; }
+    }
+}
