+++ Backend/Backend.Models/UserModels/UserRegistrationModel.cs
@@ -4,10 +4,12 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Backend.Models
+namespace Backend.Models.UserModels
 {
-    public class UserOnlyEmailModel
+    public class UserRegistrationModel
     {
         public string Email { get; set; }
+        public string Name { get; set; }
+        public string Password { get; set; }
     }
 }
