+++ Backend/Backend/Backend.csproj
@@ -182,6 +182,7 @@
   <ItemGroup>
     <Compile Include="ApplicationIdentityContext.cs" />
     <Compile Include="Controllers\AccountController.cs" />
+    <Compile Include="Controllers\ProjectController.cs" />
     <Compile Include="EnsureAuthIndexes.cs" />
     <Compile Include="IdentityConfig.cs" />
     <Compile Include="Program.cs" />
