+++ Backend/Backend/IdentityConfig.cs
@@ -7,6 +7,7 @@ using Microsoft.AspNet.Identity;
 using Microsoft.AspNet.Identity.Owin;
 using Microsoft.Owin;
 using AspNet.Identity.MongoDB;
+using Backend.Models.UserModels;
 
 namespace Backend
 {
